adding tasks & removing tasks handled

+++ views/template.php
@@ -15,12 +15,13 @@
         <div class="title">Dashboard</div>
         <div class="userPanel">
           <i class="fa fa-chevron-down"></i
-          ><span class="username">John Doe </span
-          ><img
+          ><span class="username">Welcome</span
+          > <!-- <img
             src="https://s3.amazonaws.com/uifaces/faces/twitter/kolage/73.jpg"
             width="40"
             height="40"
           />
+          -->
         </div>
       </div>
       <div class="main">
@@ -34,18 +35,20 @@
           <div class="menu">
             <div class="title">Folders</div>
             <ul class="folders">
+
+            <li class="<?= isset($_GET['folder_id']) && is_numeric($_GET['folder_id'])?'':'active'?>"><i class="fa fa-folder"></i>ALL</li>
               <?php foreach ($folders as $folder):?>  
 
-            <li>
+            <li class="<?= $folder->id == $_GET['folder_id']? 'active':'' ?>" >
             
               <a href="?folder_id=<?=  $folder->id?>">
               <i class="fa fa-folder"></i><?= $folder->name;?></a>
-              <a href="?delete_folder=<?=  $folder->id?>" class="remove">
+              <a href="?delete_folder=<?=  $folder->id?>" class="remove" onclick="return confirm('Are you sure you want to delete the item?')">
               x</a>
             </li>
               <?php endforeach;?>
-
-              <li class="active"><i class="fa fa-folder"></i>Current Folder</li>
+            
+            
               
               <div class="folder-parent">
               <input id="addFolderInput" type="text" placeholder="Add New folder" style="width: 65%;margin-left:3%;"/>
@@ -56,49 +59,40 @@
         </div>
         <div class="view">
           <div class="viewHeader">
-            <div class="title">Manage Tasks</div>
+            <div class="title">
+              <input id="addTaskInput" type="text" placeholder="Add New Task" style="width: 65%;margin-left:3%;">
+            </div>
             <div class="functions">
               <div class="button active">Add New Task</div>
               <div class="button">Completed</div>
-              <div class="button inverz"><i class="fa fa-trash-o"></i></div>
+              
             </div>
           </div>
           <div class="content">
             <div class="list">
               <div class="title">Today</div>
-              <ul>
-                <li class="checked">
-                  <i class="fa fa-check-square-o"></i
-                  ><span>Update team page</span>
+              <ul class="tasks">
+                <?php if (sizeof($tasks)>0):?>
+                <?php foreach ($tasks as $task):?>
+                <li class="<?= $task->is_done? 'checked':'' ?>">
+                  <i class="fa <?= $task->is_done? "fa-check-square-o" : "fa-square-o"?>">
+                    </i
+                  ><span><?=$task->title?></span>
+                  
                   <div class="info">
-                    <div class="button green">In progress</div>
-                    <span>Complete by 25/04/2014</span>
+                    <span class="created-at">Created_at At <?=$task->created_at?> </span>
+                    <a href="?delete_task=<?php echo $task->id; ?>" class="remove" >x</a>
                   </div>
                 </li>
-                <li>
-                  <i class="fa fa-square-o"></i><span>Design a new logo</span>
-                  <div class="info">
-                    <div class="button">Pending</div>
-                    <span>Complete by 10/04/2014</span>
-                  </div>
-                </li>
-                <li>
-                  <i class="fa fa-square-o"></i
-                  ><span>Find a front end developer</span>
-                  <div class="info"></div>
-                </li>
+                <?php endforeach; ?>
+                <?php else:?>
+                  <li>
+                    No Task here...
+                  </li>
+                <?php endif;?>
               </ul>
             </div>
-            <div class="list">
-              <div class="title">Tomorrow</div>
-              <ul>
-                <li>
-                  <i class="fa fa-square-o"></i
-                  ><span>Find front end developer</span>
-                  <div class="info"></div>
-                </li>
-              </ul>
-            </div>
+            
           </div>
         </div>
       </div>
@@ -106,6 +100,29 @@
     <!-- partial -->
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
     <script src="assets/js/script.js"></script>
-   
+    <script>
+    (
+      $(document).ready(function(){
+        $('#addTaskInput').on('keypress',function(e) {
+          e.stopPropagation();
+          if(e.which == 13) {
+            const input = $('#addTaskInput')
+            $.ajax({
+              url:'process/ajaxHandler.php',
+              method:'POST',
+              data:{
+                action:'addTask',folderId:<?=$_GET['folder_id'] ?>,taskTitle:input.val()
+              },
+              success: function(response){
+                location.reload()
+              }
+            })
+          }
+          
+        });
+        $('#addTaskInput').focus()
+      })
+    )
+    </script>
   </body>
 </html>
